backend/requirements.txt
# FastAPI and related
fastapi>=0.110.0,<0.116.0
uvicorn[standard]>=0.24.0,<0.25.0
python-multipart>=0.0.9
pydantic>=2.7.4,<3.0.0  # Specify version compatible with langchain
pydantic-settings>=2.0.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# LLM related
langchain>=0.3.0,<0.4.0  # Specific version range
langchain-community>=0.0.3

# HTTP client
requests>=2.31.0
httpx>=0.24.0

# Utilities
python-dotenv>=1.0.0